backend/app/requirements.txt
fastapi
uvicorn
openai
pyodbc
python-dotenv